# Cap Android persistent workers' max instances at 1 for optimal performance and memory usage