The global default value for `--worker_max_instances` is 4. This change turns the default number of instances for Android-related persistent workers down to 1. See https://github.com/bazelbuild/bazel/issues/8586#issuecomment-504638339 for context and benchmarks.

This change reduces the risks of Android builds OOMing when using persistent workers, and makes the builds faster as well.

tl;dr: 10-20% faster builds, 55-65% lower memory usage.

Benchmark project: https://github.com/jin/simple_tree
Heap sizes: 5.47GiB -> 2.47GiB (Linux workstation), 2.25GiB -> 811MiB (MacBook Pro)
Build wall time: 30.998s -> 24.062s (Linux workstation), 75.949s -> 67.508s (MacBook Pro)

RELNOTES: Renamed `--persistent_android_resource_processor` to `--android_persistent_workers`. Reworked the flag's internals to make Android builds up to 20% faster and reduce JVM memory usage by up to 65%.